requirements.txt
PIL==11.0.0
click==8.1.7
h5py==3.12.1
matplotlib==3.8.0
numpy==2.2.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
PyYAML==6.0.2
scipy==1.13.1
session_info==1.0.0
torch==2.5.1+cu121
torchmetrics==1.6.0
torchvision==0.20.1+cu121
tqdm==4.66.6
wandb==0.19.0
IPython==7.34.0
jupyter_client==6.1.12
jupyter_core==5.7.2
notebook==6.5.5